requirements.txt
google-api-python-client==1.5.1
googleads==4.1.0
httplib2==0.9.2
# Specific version of protobuf required to avoid ImportError on google.appengine module
# See more: https://github.com/GoogleCloudPlatform/gcloud-python/issues/1570#issuecomment-192055672
protobuf==3.0.0b2
suds-jurko==0.6
webapp2==2.5.2

# Test-only dependencies
# mock==2.0.0
# WebTest==2.0.23
